requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.27.0
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
numpy==1.26.3
pandas==2.1.4
lxml==5.0.0